Terminate IPsec connections affected by VFP VRF change.
When the VRF associated with a VFP is changed, the associated
connections must be reset. This ensures that the relevant policies
Terminate IPsec connections affected by VFP VRF change.

When the VRF associated with a VFP is changed, the associated
connections must be reset. This ensures that the relevant policies
are withdrawn. The IKE daemon will subsqquently bring them up again
and the policies will be reapplied for the correct VRF.

VRVDR-48055