vyatta-security-vpn-secrets: code injection
Avoid risk of code injection due to crl file argument passed to openssl
when executing a shell command. Handle this as a literal string by
enclosing it in single quotes, which stops any shell expansion. Note
that this vulnerability was only hit by commenting out a prior check
in this script for the existence of the crl file. Pattern restriction
for the string is not added in Yang at present, as it breaks backwards
compatibility.

VRVDR-45065