vyatta-security-vpn: code injection due to call to check_file_in_config
Avoid risk of code injection due to syntax check for file being in
/config/auth. Fix this by enclosing the variable array passed as an
argument to check_file_in_config in single quotes. This preserves the
literal value of all characters and stops interpretation by the shell
of metacharacters. Pattern restriction for these strings is not added
in Yang at present, as it breaks backwards compatibility.

VRVDR-45066