vyatta-security-vpn: set security vpn ipsec site-to-site peer code injection
Avoid risk of code injection due to the peer name (connection id)
passed when calling ipsec stroke via system cmd. Handle this as a
literal string by enclosing it in single quotes, which stops any
shell expansion for the cmd. Do the same when passing the peer name
as a variable array in config validation scripts. While most
characters are permissible in the string, pattern restriction for the
remaining few is not added in Yang at present, as it breaks backwards
compatibility.

VRVDR-45067