vyatta-security-vpn: s2s tunnel protocol code injection
Avoid risk of code injection due to syntax check for s2s tunnel when
the protocol is invalid. Handle this as a literal string by enclosing
the variable array used to pass the protocol in single quotes. There
is another instance of the variable array correctly enclosed in single
quotes. However, this still suffers from shell expansion as the cmd
is unnecessarily executed twice due to exec and use of back-ticks
within this. So simplify the script to the usual format for syntax
scripts. Use of enum for the string is not added in Yang at present,
as it breaks backwards compatibility.

VRVDR-45068