vyatta-security-vpn: set security vpn rsa-keys local-key file code injection
Avoid risk of code injection due to the input and output files for a
system call using the local key file. Handle these as literal strings by
enclosing them in single quotes, which stops any shell expansion.
Similarly, in the validation script for the local key file, enclose the
variable array for this in single quotes. Pattern restriction for the
string is not added in Yang at present, as it breaks backwards
compatibility.

VRVDR-45069