vyatta-security-vpn: l2tp remote-access dhcp-interface code injection
Avoid risk of code injection due to the inclusion of this interface
name in a sysctl system cmd. Handle this as a literal string by
enclosing it in single quotes, which stops any shell expansion. Note
that this vulnerability was only hit by commenting out prior code in
this script to substitute '.' by '/'. Restriction of the interface to
available ones is not added in Yang at present, as it breaks backwards
compatibility.

VRVDR-45071