Commit failed in IPsec site-to-site configuration
The commit failure is due to validation that an esp-group is specified
for peer tunnels. This bug is a regression due to 910f5c2b0921
("vyatta-security-vpn: set security vpn ipsec site-to-site peer code
injection"). The problem is due to the addition of quotes for the
tunnel id. This results in the failure of the check as this is a uint32
and not a string, so there is no risk of code injection here anyway.

VRVDR-49924