Avoid setup-race between VCI and non-VCI code
If the VCI IPsec component start it also starts the strongswan
service, which loads the non-VCI legacy configuration file for
non-VCI aware features. This causes non-VCI tunnel configuration
gets loaded and initiated from the previous boot.

This has also the downside that strongswan gets started before
vyatta-ike-sa-daemon is up and running and misses the updown
events of the old-configuration tunnel. The non-VCI code
tears down the pre-boot tunnel. Since vyatta-ike-sa-daemon
missed the updown event, it will not be able to bring this
tunnel up again.

Original sympthom: site-to-site tunnel spuriosly not come
up on a reboot of an installed image with an existing
site-to-site configuration.

VRVDR-50831