vpn rsa-keys local-key: do not enforce that RSA private key file exists
The handling for generate vpn rsa-key is changed so that the RSA key
file cannot be overwritten, where its path is specified by the vpn
rsa-keys local-key configuration. The latter mandates that the file
already exists. A workaround to allow key generation in a non-default
location is to first create an empty key file. This is no longer
possible, so the solution is to no longer enforce the existence of the
key file for the configuration.

VRVDR-52388

(cherry picked from commit a803816870b19b478de60e526e499996beab9d92)